Fix a warning in the `simple` example.

+++ examples/crossBuilds/simple/library/js/src/main/scala/simple/Platform.scala
@@ -4,6 +4,6 @@ import scalajs.js
 
 object Platform{
   def format(ts: Long) = {
-    new js.Date(ts).toISOString()
+    new js.Date(ts.toDouble).toISOString()
   }
 }
